requirements.txt
azure-ai-projects>=2.0.0b1
azure-identity>=1.19.0
